pyproject.toml
[project]
name = "help-center-rag"
version = "0.1.0"
description = "Generic Help Center Documentation RAG System using Anthropic's Contextual Retrieval methodology"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  # HTML to Markdown conversion
  "beautifulsoup4>=4.13.3",
  "html2text>=2024.2.26",
  "markdown-it-py>=3.0.0",
  "markdownify>=1.1.0",
  # RAG System
  "aiosqlite>=0.21.0",
  "asyncio>=3.4.3",
  "bm25s>=0.2.9",
  "einops>=0.8.1",
  "flash-attn>=2.7.4.post1",
  "google-generativeai>=0.8.4",
  "httpx>=0.28.1",
  "litellm>=1.63.11",
  "loguru>=0.7.3",
  "openai>=1.66.3",
  "pillow>=11.1.0",
  "pydantic-settings>=2.8.1",
  "pydantic>=2.8.2",
  "qdrant-client>=1.13.3",
  "rank-bm25>=0.2.2",
  "rich>=13.9.4",
  "sentence-transformers>=3.4.1",
  "tiktoken>=0.9.0",
  "torch>=2.6.0",
  "transformers>=4.49.0",
  "typer>=0.15.2",
  "zenml[server]>=0.80",
  "pyright>=1.1.397",
  "scikit-learn>=1.6.1",
  "matplotlib>=3.10.3",
  "seaborn>=0.13.2",
  "numpy>=2.2.4",
  "pandas>=2.3.0",
]

[dependency-groups]
dev = [
  "black>=25.1.0",
  "flake8>=7.1.2",
  "pyright>=1.1.396",
  "pytest-asyncio>=0.21.0",
  "pytest>=7.3.1",
  "ruff>=0.11.0",
  "autoflake>=2.3.1",
]

# [project.scripts]
# help-rag = "help_rag.interfaces.cli.main:app"
